requirements.txt
streamlit
langchain_core
langchain
langchain-google-genai
python-dotenv
pydantic
